perf: make conflation explicit and manual
This is necessary to provide clients with the correct information when the request is discarded.

Discarding is performed when too many requests arrive and just the last one is kept, maintaining alive the last running request. For each discarded element, we send the appropriate message.

Manual conflation is necessary because there isn't a direct and fast approach to register a callback for elements dropped in a conflated channel. For this reason that operation has been made explicit in order to provide consumers with a more meaningful result instead of an empty completion list.