feat: use `WebFlux`-based web server for lsp-based completions
This commit introduces a new dedicated module for completions. This module leverages a `WebFlux` web-server for its springboot application in order to exploit reactive patterns both in the LSP RESTful approach and WebSocket. This helps in preventing clashes with default compiler application which uses Servlet and forced us to use an old implementation of WebSockets (`TextWebSocketHandler`).